requirements-workflows.txt
# Repazoo Workflows Dependencies
# Prefect 2.x orchestration with compliance features

# Core Workflow Engine
prefect==2.14.10
prefect-docker==0.4.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis (for rate limiting)
redis==5.0.1
hiredis==2.2.3

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Encryption
cryptography==41.0.7

# Data Processing
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
